use broadcast channels instead of notify/wait mechanism for shutdown
The notify / wait mechanism worked *most of the time*, but if a task
was not currently awaiting a notification while it was sent, it would
never receive it once it started waiting for it again. For example,
if a task was busy when the shutdown signal was received, it would
never receive the notification, and block clean shutdown.

Hence, change the mechanism to broadcast channels, which will always
yield a message to any subscribed receiver, even if it wasn't
actively listening when the message was sent.