use larger bufsize for WireGuard and drop packets that are too big
Tunn::encapsulate panics when it's called with a payload that is too
large for the target buffer. This commit makes the buffer for
WireGuard packets a lot bigger (1500 -> 65535 bytes), and takes the
size of WireGuard packet headers into account.

Packets for which the payload will not fit into the buffer are now
dropped instead of crashing the WireGuard server.

This should fix problems when attempting to send UDP packets with
huge payloads back over the WireGuard tunnel.

Resolves #17